refactor(lessons/repositories): alteraçoes nos imports e exports
