+++ src/modules/lessons/repositories/IChannels.repository.ts
@@ -1,9 +1,7 @@
-import { CreateChannelDto } from '@lessons:dtos/CreateChannel.dto';
-import Channel from '@lessons:entities/Channel';
+import { CreateChannelDto } from '@lessons:dtos/channels/CreateChannel.dto';
+import { Channel } from '@lessons:entities/Channel';
 
-interface IChannelRepository {
+export interface IChannelsRepository {
   create(data: CreateChannelDto): Promise<Channel>;
   findById(id: string): Promise<Channel | undefined>;
 }
-
-export default IChannelRepository;
