+++ src/modules/lessons/repositories/ISubjects.repository.ts
@@ -1,10 +1,8 @@
-import { CreateSubjectDto } from '@lessons:dtos/CreateSubject.dto';
-import Subject from '@lessons:entities/Subject';
+import { CreateSubjectDto } from '@lessons:dtos/subject/CreateSubject.dto';
+import { Subject } from '@lessons:entities/Subject';
 import { Uuid } from '@shared/@types/Uuid';
 
-interface ISubjectsRepository {
+export interface ISubjectsRepository {
   create(data: CreateSubjectDto): Promise<Subject>;
   findById(id: Uuid): Promise<Subject | undefined>;
 }
-
-export default ISubjectsRepository;
