+++ src/modules/lessons/repositories/IVideos.repository.ts
@@ -1,13 +1,11 @@
-import { CreateVideoDto } from '@lessons:dtos/CreateVideo.dto';
-import Video from '@lessons:entities/Video';
+import { CreateVideoDto } from '@lessons:dtos/videos/CreateVideo.dto';
+import { Video } from '@lessons:entities/Video';
 
 // ---------------------------------------------------------------------------------------------- //
 
-interface IVideosRepository {
+export interface IVideosRepository {
   create(data: CreateVideoDto): Promise<Video>;
   findById(id: string): Promise<Video | undefined>;
 }
 
 // ---------------------------------------------------------------------------------------------- //
-
-export default IVideosRepository;
